refactor: move BindItemsTest tests into SignalTest classes (#8911)
Consolidate bind-items exception tests from 7 separate *BindItemsTest
classes into the *SignalTest pattern, which is the standard home for
signal-related unit tests. Uses private helper methods instead of @Before
to keep the classes extensible for future signal tests.
