+++ vaadin-grid-flow-parent/vaadin-grid-flow/src/test/java/com/vaadin/flow/component/grid/GridSignalTest.java
@@ -13,67 +13,65 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.flow.component.listbox;
+package com.vaadin.flow.component.grid;
 
 import java.util.Arrays;
 
-import org.junit.Before;
-import org.junit.Rule;
 import org.junit.Test;
 
 import com.vaadin.flow.data.provider.DataProvider;
 import com.vaadin.flow.data.provider.ListDataProvider;
 import com.vaadin.flow.signals.BindingActiveException;
 import com.vaadin.flow.signals.local.ListSignal;
-import com.vaadin.tests.MockUIRule;
+import com.vaadin.tests.AbstractSignalsUnitTest;
 
-public class ListBoxBindItemsTest {
-
-    @Rule
-    public final MockUIRule ui = new MockUIRule();
-
-    private ListBox<String> listBox;
-
-    @Before
-    public void setup() {
-        listBox = new ListBox<>();
-        var itemsSignal = new ListSignal<String>();
-        itemsSignal.insertLast("Item 1");
-        itemsSignal.insertLast("Item 2");
-        listBox.bindItems(itemsSignal);
-        ui.getUI().add(listBox);
-    }
+public class GridSignalTest extends AbstractSignalsUnitTest {
 
     @Test(expected = BindingActiveException.class)
     public void bindItems_thenSetDataProvider_throws() {
-        listBox.setDataProvider(
-                DataProvider.ofItems("New Item 1", "New Item 2"));
+        var grid = createGridWithBoundItems();
+        grid.setDataProvider(DataProvider.ofItems("New Item 1", "New Item 2"));
     }
 
     @Test(expected = BindingActiveException.class)
     public void bindItems_thenSetItemsWithDataProvider_throws() {
-        listBox.setItems(DataProvider.ofItems("New Item 1", "New Item 2"));
+        var grid = createGridWithBoundItems();
+        grid.setItems(DataProvider.ofItems("New Item 1", "New Item 2"));
     }
 
     @Test(expected = BindingActiveException.class)
     public void bindItems_thenSetItemsWithInMemoryDataProvider_throws() {
-        listBox.setItems(DataProvider
+        var grid = createGridWithBoundItems();
+        grid.setItems(DataProvider
                 .ofCollection(Arrays.asList("New Item 1", "New Item 2")));
     }
 
     @Test(expected = BindingActiveException.class)
     public void bindItems_thenSetItemsWithListDataProvider_throws() {
-        listBox.setItems(new ListDataProvider<>(
+        var grid = createGridWithBoundItems();
+        grid.setItems(new ListDataProvider<>(
                 Arrays.asList("New Item 1", "New Item 2")));
     }
 
     @Test(expected = BindingActiveException.class)
     public void bindItems_thenSetItemsWithCollection_throws() {
-        listBox.setItems(Arrays.asList("New Item 1", "New Item 2"));
+        var grid = createGridWithBoundItems();
+        grid.setItems(Arrays.asList("New Item 1", "New Item 2"));
     }
 
     @Test(expected = BindingActiveException.class)
     public void bindItems_thenSetItemsWithVarargs_throws() {
-        listBox.setItems("New Item 1", "New Item 2");
+        var grid = createGridWithBoundItems();
+        grid.setItems("New Item 1", "New Item 2");
+    }
+
+    private Grid<String> createGridWithBoundItems() {
+        var grid = new Grid<String>();
+        var itemsSignal = new ListSignal<String>();
+        itemsSignal.insertLast("Item 1");
+        itemsSignal.insertLast("Item 2");
+        grid.bindItems(itemsSignal);
+        ui.add(grid);
+        return grid;
     }
 }
