+++ vaadin-radio-button-flow-parent/vaadin-radio-button-flow/src/test/java/com/vaadin/flow/component/radiobutton/RadioButtonGroupSignalTest.java
@@ -15,6 +15,7 @@
  */
 package com.vaadin.flow.component.radiobutton;
 
+import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -22,7 +23,10 @@ import org.junit.Assert;
 import org.junit.Test;
 
 import com.vaadin.flow.component.UI;
+import com.vaadin.flow.data.provider.DataProvider;
+import com.vaadin.flow.data.provider.ListDataProvider;
 import com.vaadin.flow.signals.BindingActiveException;
+import com.vaadin.flow.signals.local.ListSignal;
 import com.vaadin.flow.signals.local.ValueSignal;
 import com.vaadin.tests.AbstractSignalsUnitTest;
 
@@ -86,10 +90,60 @@ public class RadioButtonGroupSignalTest extends AbstractSignalsUnitTest {
                 buttons.get(2).isEnabled());
     }
 
+    @Test(expected = BindingActiveException.class)
+    public void bindItems_thenSetDataProvider_throws() {
+        var radioButtonGroup = createRadioButtonGroupWithBoundItems();
+        radioButtonGroup.setDataProvider(
+                DataProvider.ofItems("New Item 1", "New Item 2"));
+    }
+
+    @Test(expected = BindingActiveException.class)
+    public void bindItems_thenSetItemsWithDataProvider_throws() {
+        var radioButtonGroup = createRadioButtonGroupWithBoundItems();
+        radioButtonGroup
+                .setItems(DataProvider.ofItems("New Item 1", "New Item 2"));
+    }
+
+    @Test(expected = BindingActiveException.class)
+    public void bindItems_thenSetItemsWithInMemoryDataProvider_throws() {
+        var radioButtonGroup = createRadioButtonGroupWithBoundItems();
+        radioButtonGroup.setItems(DataProvider
+                .ofCollection(Arrays.asList("New Item 1", "New Item 2")));
+    }
+
+    @Test(expected = BindingActiveException.class)
+    public void bindItems_thenSetItemsWithListDataProvider_throws() {
+        var radioButtonGroup = createRadioButtonGroupWithBoundItems();
+        radioButtonGroup.setItems(new ListDataProvider<>(
+                Arrays.asList("New Item 1", "New Item 2")));
+    }
+
+    @Test(expected = BindingActiveException.class)
+    public void bindItems_thenSetItemsWithCollection_throws() {
+        var radioButtonGroup = createRadioButtonGroupWithBoundItems();
+        radioButtonGroup.setItems(Arrays.asList("New Item 1", "New Item 2"));
+    }
+
+    @Test(expected = BindingActiveException.class)
+    public void bindItems_thenSetItemsWithVarargs_throws() {
+        var radioButtonGroup = createRadioButtonGroupWithBoundItems();
+        radioButtonGroup.setItems("New Item 1", "New Item 2");
+    }
+
     @SuppressWarnings("unchecked")
     private List<RadioButton<String>> getRadioButtons() {
         return group.getChildren().filter(RadioButton.class::isInstance)
                 .map(child -> (RadioButton<String>) child)
                 .collect(Collectors.toList());
+    }
+
+    private RadioButtonGroup<String> createRadioButtonGroupWithBoundItems() {
+        var radioButtonGroup = new RadioButtonGroup<String>();
+        var itemsSignal = new ListSignal<String>();
+        itemsSignal.insertLast("Item 1");
+        itemsSignal.insertLast("Item 2");
+        radioButtonGroup.bindItems(itemsSignal);
+        ui.add(radioButtonGroup);
+        return radioButtonGroup;
     }
 }
