--- vaadin-select-flow-parent/vaadin-select-flow/src/test/java/com/vaadin/flow/component/select/SelectBindItemsTest.java
@@ -1,80 +1,0 @@
-/*
- * Copyright 2000-2026 Vaadin Ltd.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not
- * use this file except in compliance with the License. You may obtain a copy of
- * the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations under
- * the License.
- */
-package com.vaadin.flow.component.select;
-
-import java.util.Arrays;
-
-import org.junit.Before;
-import org.junit.Rule;
-import org.junit.Test;
-
-import com.vaadin.flow.data.provider.DataProvider;
-import com.vaadin.flow.data.provider.ListDataProvider;
-import com.vaadin.flow.signals.BindingActiveException;
-import com.vaadin.flow.signals.local.ListSignal;
-import com.vaadin.tests.MockUIRule;
-
-public class SelectBindItemsTest {
-
-    @Rule
-    public final MockUIRule ui = new MockUIRule();
-
-    private Select<String> select;
-
-    @Before
-    public void setup() {
-        select = new Select<>();
-        var itemsSignal = new ListSignal<String>();
-        itemsSignal.insertLast("Item 1");
-        itemsSignal.insertLast("Item 2");
-        select.bindItems(itemsSignal);
-        ui.getUI().add(select);
-    }
-
-    @Test(expected = BindingActiveException.class)
-    public void bindItems_thenSetDataProvider_throws() {
-        select.setDataProvider(
-                DataProvider.ofItems("New Item 1", "New Item 2"));
-    }
-
-    @Test(expected = BindingActiveException.class)
-    public void bindItems_thenSetItemsWithDataProvider_throws() {
-        select.setItems(DataProvider.ofItems("New Item 1", "New Item 2"));
-    }
-
-    @Test(expected = BindingActiveException.class)
-    public void bindItems_thenSetItemsWithInMemoryDataProvider_throws() {
-        select.setItems(DataProvider
-                .ofCollection(Arrays.asList("New Item 1", "New Item 2")));
-    }
-
-    @Test(expected = BindingActiveException.class)
-    public void bindItems_thenSetItemsWithListDataProvider_throws() {
-        select.setItems(new ListDataProvider<>(
-                Arrays.asList("New Item 1", "New Item 2")));
-    }
-
-    @Test(expected = BindingActiveException.class)
-    public void bindItems_thenSetItemsWithCollection_throws() {
-        select.setItems(Arrays.asList("New Item 1", "New Item 2"));
-    }
-
-    @Test(expected = BindingActiveException.class)
-    public void bindItems_thenSetItemsWithVarargs_throws() {
-        select.setItems("New Item 1", "New Item 2");
-    }
-
-}
