+++ vaadin-virtual-list-flow-parent/vaadin-virtual-list-flow/src/test/java/com/vaadin/flow/component/virtuallist/tests/VirtualListSignalTest.java
@@ -18,50 +18,47 @@ package com.vaadin.flow.component.virtuallist.tests;
 import java.util.Collections;
 import java.util.stream.Stream;
 
-import org.junit.Before;
-import org.junit.Rule;
 import org.junit.Test;
 
 import com.vaadin.flow.component.virtuallist.VirtualList;
 import com.vaadin.flow.data.provider.DataProvider;
 import com.vaadin.flow.signals.BindingActiveException;
 import com.vaadin.flow.signals.local.ListSignal;
-import com.vaadin.tests.MockUIRule;
+import com.vaadin.tests.AbstractSignalsUnitTest;
 
-public class VirtualListBindItemsTest {
-
-    @Rule
-    public final MockUIRule ui = new MockUIRule();
-
-    private VirtualList<String> list;
-
-    @Before
-    public void setup() {
-        list = new VirtualList<>();
-        var itemsSignal = new ListSignal<String>();
-        itemsSignal.insertLast("Item 1");
-        itemsSignal.insertLast("Item 2");
-        list.bindItems(itemsSignal);
-        ui.getUI().add(list);
-    }
+public class VirtualListSignalTest extends AbstractSignalsUnitTest {
 
     @Test(expected = BindingActiveException.class)
     public void bindItems_thenSetDataProvider_throws() {
+        var list = createVirtualListWithBoundItems();
         list.setDataProvider(DataProvider.ofItems("New Item 1", "New Item 2"));
     }
 
     @Test(expected = BindingActiveException.class)
     public void bindItems_thenSetCollection_throws() {
+        var list = createVirtualListWithBoundItems();
         list.setItems(Collections.emptyList());
     }
 
     @Test(expected = BindingActiveException.class)
     public void bindItems_thenSetItems_throws() {
+        var list = createVirtualListWithBoundItems();
         list.setItems("New Item 1");
     }
 
     @Test(expected = BindingActiveException.class)
     public void bindItems_thenSetStream_throws() {
+        var list = createVirtualListWithBoundItems();
         list.setItems(Stream.of("New Item 1", "New Item 2"));
+    }
+
+    private VirtualList<String> createVirtualListWithBoundItems() {
+        var list = new VirtualList<String>();
+        var itemsSignal = new ListSignal<String>();
+        itemsSignal.insertLast("Item 1");
+        itemsSignal.insertLast("Item 2");
+        list.bindItems(itemsSignal);
+        ui.add(list);
+        return list;
     }
 }
